Save pages more selectively and conserve history on panel close
This fixes the unwanted behavior where Search in Sidebar would save pages that were navigated to via extensions, or via the main browser window.

This also makes it so the history is conserved when closing the panel, so you can quickly go back to where you left off by clicking the extension icon to open the panel again. The navigation history is now only cleared when making a new search via the context menu, which feel much more logical to me.